% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMK.R
\name{AMK}
\alias{AMK}
\title{Additive Multiplicative Kernel Regression}
\usage{
AMK(
  trainX,
  trainY,
  testX,
  bw = "dpi",
  nMultiCov = 3,
  fixedCov = c(1, 2),
  cirCov = 2
)
}
\arguments{
\item{cirCov}{}
}
\description{
Additive Multiplicative Kernel Regression
}
