% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePCurve.R
\name{ComputePCurve}
\alias{ComputePCurve}
\title{Power curve comparison}
\usage{
ComputePCurve(
  data,
  xcol,
  xcol.circ = NULL,
  ycol,
  var1col,
  var2col,
  thrs = 0.2,
  conflevel = 0.95,
  gridSize = 50
)
}
\arguments{
\item{data}{a List of data sets to be compared}

\item{xcol}{a numeric or vector stating column number of covariates}

\item{xcol.circ}{a numeric or vector stating column number of circular covariates}

\item{ycol}{A numeric value stating the column number of target}

\item{var1col}{a numeric value stating column number of first covariate to used in generating test set}

\item{var2col}{a numeric value stating column number of second covariate to be used in generating test set}

\item{thrs}{A single value or vector represnting threshold for each covariates}

\item{conflevel}{A single value as a Bound to be used in hypothesis testing while function comparison}

\item{gridSize}{A single value to used in constructing test set size}
}
\value{
a list containing :
 \itemize{
  \item muDiff - The testset prediction for each of the data set
  \item mu1 - The test prediction for first data set
  \item mu2 - The test prediction for second data set
  \item band - The allowed statistical difference between functions
  \item confLevel - The statistical boundation
  \item testset - The test prediction for first data set
  \item estimatedParams - The function parameter values
  \item statisticalDiff - The \% statistical difference between functions in comaprison
  \item weightedDiff - The \% wighted difference between functions in comparison
  \item ratioVarcol1 - The shrinkage ratio of varcol1
  \item ratioVarcol1 - The shrinkage ratio of varcol2
  \item extrapolatedPower - The matched power difference scaled on entire domain of a covariate
}
}
\description{
Power curve comparison
}
