% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatch.R
\name{covMatch}
\alias{covMatch}
\title{Covariate Matching}
\usage{
covMatch(dname, xCol, xCol.circ, wgt, priority)
}
\arguments{
\item{data}{a list, consisting of data sets to match, also each of the individual data set can be dataframe or a matrix}

\item{xCol}{a vector stating the column position of covariates used.}

\item{xCol.circ}{a vector stating the column position of circular variables}

\item{wgt}{a numerical or a vector of threshold values for each covariates, against which matching happens
It should be a single value or a vector of values representing threshold for each of the covariate}
}
\value{
a list containing :
  \itemize{
  \item originalData - The data sets provided for matching
  \item matchedData - The data sets after matching
  \item MinMaxOriginal - The minimum and maximum value in original data for each covariate used in matching
  \item MinMaxMatched - The minimum and maximum value in matched data for each covariates used in matching
}
}
\description{
The function aims to take list of two data sets and returns the after
matched data sets using user specified covariates and threshold
}
