% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Wind Energy data set containing 47,542 data points}
\format{
A data frame with 47,542 rows and 7 variables
}
\usage{
data(data1)
}
\description{
A dataset containing the power produced and other attributes of almost 47,542
records.
}
\details{
\itemize{
  \item Data.point - sequence of integers displaying each record
  \item V - wind speed
  \item D - wind direction
  \item air.density - air density
  \item I - turbulence intensity
  \item S_b - wind shear
  \item Y - wind power
}
}
\keyword{datasets}
