% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGP.R
\name{funGP}
\alias{funGP}
\title{Function estimator using Gaussian Process}
\usage{
funGP(datalist, covCols, yCol, confLevel, testset)
}
\arguments{
\item{datalist}{List of data sets}

\item{covCols}{Vector of column number of covariates}

\item{yCol}{A numeric value for target}

\item{confLevel}{Statistical confidence value, such as 0.95}

\item{testset}{Test points at which the functions will be compared}
}
\description{
Function estimator using Gaussian Process
}
