% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnPCurve.R
\name{KnnPcurve}
\alias{KnnPcurve}
\title{KNN : PowerCurve}
\usage{
KnnPcurve(dname, xvar, xvar.circ = NULL, yvar, kfold, forward_stepwise = FALSE)
}
\arguments{
\item{dname}{The data set to be used for modelling, could be a dataframe or matrix}

\item{xvar}{a vector with column number of non circular features}

\item{xvar.circ}{a vector with column number of circular features}

\item{yvar}{column number of target}

\item{kfold}{split the data set should go through, such as kfold = 10}

\item{forward_stepwise}{variable selection technique, which gives the best result for 1 feature, 2 features etc}
}
\value{

}
\description{
The function models the powercurve using KNN, against supplied arguments
}
