% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnPredict.R
\name{KnnPredict}
\alias{KnnPredict}
\title{KNN : Predict}
\usage{
KnnPredict(knnMdl, testData)
}
\arguments{
\item{testData}{a data frame or matrix, to compute the predictions}

\item{KnnMdl}{a list containing:
1) data
2) column number of features
3) column number of target
4) bestK from trained model}
}
\value{
a list containing :
        1) data
        2) column number of features
        3) column number of target
        4) prediction on test data
}
\description{
The function can be used to make prediction on test data using trained model
}
