% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnUpdate.R
\name{KnnUpdate}
\alias{KnnUpdate}
\title{KNN : Update}
\usage{
KnnUpdate(knnMdl, newData)
}
\arguments{
\item{newData}{a dataframe or a matrix, to be used for updating the model}

\item{KnnMdl}{a list containing:
1) data
2) column number of features
3) column number of target
4) bestK from trained model}
}
\value{
a list containing :
        1) data
        2) column number of features
        3) column number of target
        4) best K
}
\description{
The function can be used to update KNN model when new data is provided
}
