% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempGP.R
\name{predict.tempGP}
\alias{predict.tempGP}
\title{predict from temporal Gaussian process}
\usage{
\method{predict}{tempGP}(tempGPObj, testX)
}
\arguments{
\item{tempGPObj}{An object of class tempGP}

\item{testX}{A matrix with each column corresponding to one input variable.}
}
\value{
a vector of predictions at the testpoints in \code{testX}.
}
\description{
predict function for tempGP objects
}
