% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempGP.R
\name{updateData}
\alias{updateData}
\title{Updating data in a model}
\usage{
updateData(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{additional arguments for passing to specific methods}
}
\value{
The returned value would depend on the class of its argument \code{object}.
}
\description{
\code{updateData} is a generic function to update data in a model.
}
\seealso{
\code{\link{updateData.tempGP}}
}
