% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempGP.R
\name{updateData.tempGP}
\alias{updateData.tempGP}
\title{Update the data in a tempGP object}
\usage{
\method{updateData}{tempGP}(
  tempGPObj,
  newX,
  newY,
  newT = NULL,
  replace = TRUE,
  updateModelF = FALSE
)
}
\arguments{
\item{tempGPObj}{An object of class tempGP.}

\item{newX}{A matrix with each column corresponding to one input variable.}

\item{newY}{A vector with each element corresponding to the output at the corresponding row of \code{newX}.}

\item{newT}{A vector with time indices of the new datapoints. If \code{NULL}, the function assigns natural numbers starting with one larger than the existing time indices in \code{trainT}.}

\item{replace}{A boolean to specify whether to replace the old data with the new one, or to add the new data while still keeping all the old data. Default is TRUE, which replaces the top \code{m} rows from the old data, where \code{m} is the number of data points in the new data.}

\item{updateModelF}{A boolean to specify whether to update \code{modelF} as well. If the original \code{tempGP} model is trained on a sufficiently large dataset (say one year), updating \code{modelF} regularly may not result in any significant improvement, but can be computationally expensive.}
}
\value{
An updated object of class \code{tempGP}.
}
\description{
This function updates \code{trainX}, \code{trainY}, and \code{trainT} in a \code{tempGP} object. By default, if the new data has \code{m} data points, the function removes top \code{m} data points from the tempGP object and appends the new data at the bottom, thus keeping the total number of data points the same. This can be overwritten by setting \code{replace = FALSE} to keep all the data points (old and new). The method also updates \code{modelG} by computing and updating residuals at the new data points. \code{modelF} can be also be updated by setting the argument \code{updateModelF} to \code{TRUE}, though not required generally (see comments in the \code{Arguments}.)
}
