% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMK.R
\name{AMK}
\alias{AMK}
\title{Additive Multiplicative Kernel Regression}
\usage{
AMK(
  trainX,
  trainY,
  testX,
  bw = "dpi",
  nMultiCov = ncol(trainX),
  fixedCov = NA,
  cirCov = NA
)
}
\arguments{
\item{trainX}{a matrix or dataframe to be used in modelling}

\item{trainY}{a numeric or vector as a target}

\item{testX}{a matrix or dataframe, to be used in computing the predictions}

\item{bw}{a vector or character input. If character, the input should be 'dpi' or 'dpi-gap'}

\item{nMultiCov}{a numerical value specifying the number of covariates in multiplicative term}

\item{fixedCov}{a vector or numeric specifying the fixed covariates column number, default value is NA}

\item{cirCov}{a vector or numeric specifying the circular covariates column number, default value is NA

' @return a list containing :
  \itemize{
  \item bandwidth - a smoothing parameter for each of the features used in modelling
  \item pred - The predictions on user provided test data
}}
}
\description{
Additive Multiplicative Kernel Regression
}
