% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMK.R
\name{AMK}
\alias{AMK}
\title{Additive Multiplicative Kernel Regression}
\usage{
AMK(
  trainX,
  trainY,
  testX,
  bw = "dpi",
  nMultiCov = ncov(trainX),
  fixedCov = NULL,
  cirCov = NULL
)
}
\arguments{
\item{trainX}{a matrix or dataframe to be used in modelling}

\item{trainY}{a numeric or vector as a target}

\item{testX}{a matrix or dataframe, to be used in computing the predictions}

\item{bw}{a vector of character input. If character, the input should be 'dpi' or 'dpi-gap'}

\item{nMultiCov}{a numerical value specifying the number of covariates in multiplicative term}

\item{fixedCov}{a vector or numeric specifying the fixed covariates column number}

\item{cirCov}{a vector or numeric specifying the circular covariates column number}
}
\description{
Additive Multiplicative Kernel Regression
}
