% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePCurve.R
\name{ComparePCurve}
\alias{ComparePCurve}
\title{Power curve comparison}
\usage{
ComparePCurve(
  data,
  xCol,
  xCol.circ = NULL,
  yCol,
  testCol,
  testSet = NULL,
  thrs = 0.2,
  conflevel = 0.95,
  gridSize = c(50, 50),
  limitMemory = T
)
}
\arguments{
\item{data}{a list of data sets to be compared}

\item{xCol}{a numeric or vector stating column number of covariates}

\item{xCol.circ}{a numeric or vector stating column number of circular covariates}

\item{yCol}{A numeric value stating the column number of target}

\item{testCol}{a numeric/vector stating column number of covariates to used in generating test set}

\item{testSet}{a matrix or dataframe consisting of test points, default value NULL, if NULL computes test points internally using testCol variables}

\item{thrs}{A single value or vector represnting threshold for each covariates}

\item{conflevel}{a single value representing the statistical significance level for constructing the band}

\item{gridSize}{A numeric / vector to be used in constructing test set, should be provided when testSet is NuLL, else it is ignored}

\item{limitMemory}{A boolean (True/False) indicating whether to limit the memory use or not. Default is true. If set to true, 5000 datapoints are randomly sampled from each dataset under comparison for inference}
}
\value{
a list containing :
 \itemize{
  \item weightedDiff - a numeric,  \% difference between the functions weighted using the density of the covariates
  \item weightedStatDiff - a numeric, \% statistically significant difference between the functions weighted using the density of the covariates
  \item scaledDiff -  a numeric, \% difference between the functions scaled to the orginal data
  \item scaledStatDiff - a numeric, \% statistically significant difference between the functions scaled to the orginal data
  \item unweightedDiff - a numeric,  \% difference between the functions unweighted
  \item unweightedStatDiff - a numeric,  \% statistically significant difference between the functions unweighted
  \item reductionRatio -  a list consisting of shrinkage ratio of features used in testSet
  \item muDiff - a vector of the difference in prediction for each test point
  \item mu1 - a vector of prediction on testset using the first data set
  \item mu2 - a vector of prediction on testset using the second data set
  \item band - a vector for the confidence band at all the testpoints for the two functions to be the same at a given cofidence level.
  \item confLevel - a numeric representing the statistical significance level for constructing the band
  \item testSet - a vector/matrix of the test points either provided by user, or generated internally
  \item estimatedParams - a list of estimated hyperaparameters for the Gaussian process model
}
}
\description{
Power curve comparison
}
