% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePCurve.R
\name{ComparePCurve}
\alias{ComparePCurve}
\title{Power curve comparison}
\usage{
ComparePCurve(
  data,
  xCol,
  xCol.circ = NULL,
  yCol,
  testCol,
  testSet = NULL,
  thrs = 0.2,
  conflevel = 0.95,
  gridSize = c(50, 50)
)
}
\arguments{
\item{data}{a list of data sets to be compared}

\item{xCol}{a numeric or vector stating column number of covariates}

\item{xCol.circ}{a numeric or vector stating column number of circular covariates}

\item{yCol}{A numeric value stating the column number of target}

\item{testCol}{a numeric/vector stating column number of covariates to used in generating test set}

\item{testSet}{a matrix or dataframe consisting of test points, default value NULL, if NULL computes test points internally using testCol variables}

\item{thrs}{A single value or vector represnting threshold for each covariates}

\item{conflevel}{a single value representing the statistical significance level for constructing the band}

\item{gridSize}{A numeric / vector to be used in constructing test set, should be provided when testSet is NuLL, else it is ignored}
}
\value{
a list containing :
 \itemize{
  \item muDiff - The testSet prediction for each of the data set
  \item mu1 - The test prediction for first data set
  \item mu2 - The test prediction for second data set
  \item band - The allowed statistical difference between functions
  \item confLevel - a single value representing the statistical significance level for constructing the band
  \item testSet - The test set provided by user, or generated internally
  \item estimatedParams - The function parameter values
  \item statDiff - The \% statistical difference between functions in comparison
  \item scaledStatDiff - The \% statistical difference scaled on entire domain of the covariates
  \item Diff - The \% difference between functions in comparison
  \item scaledDiff - The \% difference scaled on entire domain of the covariates
  \item reductionRatio - a list consisting of shrinkage ratio of features used in testSet
}
}
\description{
Power curve comparison
}
