% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePCurve.R
\name{ComputePCurve}
\alias{ComputePCurve}
\title{Power curve comparison}
\usage{
ComputePCurve(
  data,
  xcol,
  xcol.circ = NULL,
  ycol,
  var1col,
  var2col,
  thrs = 0.2,
  conflevel = 0.95,
  gridSize = 50
)
}
\arguments{
\item{data}{List of data sets to be compared}

\item{xcol}{Vector of covariates column number}

\item{xcol.circ}{vector of circular covariates column number}

\item{ycol}{A value representing column number of target}

\item{var1col}{Column number of first covariate to used in generating test set}

\item{var2col}{Column number of second covariate to be used in generating test set}

\item{thrs}{A single value or vector represnting threshold for each covariates}

\item{conflevel}{A single value as a Bound to be used in hypothesis testing while function comparison}

\item{gridSize}{A single value to used in constructing test set size}
}
\description{
Power curve comparison
}
