% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatch.R
\name{CovMatch}
\alias{CovMatch}
\title{Covariate Matching}
\usage{
CovMatch(dname, cov, weight, cov_circ)
}
\arguments{
\item{dname}{This should always be a list, containing all the data sets to match.}

\item{cov}{Vector of column numbers for the covariates to be used in matching.}

\item{weight}{Vector of threshold values, against which matching happens.
It should be a vector such as c(0.2, 0.2, 0.3), if considering three covariates.}

\item{cov_circ}{Vector stating the column position of circular variables such as wind direction,
nacelle position etc.}
}
\value{
The function returns a list containing after matched data sets.
}
\description{
The function aims to take exactly two data sets and returns the after
matched data sets using user specified covariates.
}
