% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatch.R
\name{CovMatch}
\alias{CovMatch}
\title{Covariate Matching}
\usage{
CovMatch(dname, cov, wgt, cov.circ)
}
\arguments{
\item{dname}{This should always be a list, containing data sets to match.}

\item{cov}{Vector stating the column position of non circular variables such as wind speed, ambient temperature etc.}

\item{cov.circ}{Vector stating the column position of circular variables such as wind direction,
nacelle position etc.}

\item{wgt}{Vector of threshold values for each covariates, against which matching happens.
It should be a single value (0.2) or a vector of values c(0.2, 0.3) representing threshold for each of the covariate.}
}
\value{
The function returns a list containing after matched data sets.
}
\description{
The function aims to take list of two data sets and returns the after
matched data sets using user specified covariates and threshold.
}
