% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGP.R
\name{funGP}
\alias{funGP}
\title{Function comparison using Gaussian Process and Hypothesis testing}
\usage{
funGP(
  datalist,
  xCol,
  yCol,
  confLevel = 0.95,
  testset,
  limitMemory = T,
  opt_method = "L-BFGS-B"
)
}
\arguments{
\item{datalist}{a list of data sets to compute a function for each of them}

\item{xCol}{a numeric or vector stating the column number of covariates}

\item{yCol}{A numeric value stating the column number of target}

\item{confLevel}{a single value representing the statistical significance level for constructing the band}

\item{testset}{Test points at which the functions will be compared}

\item{limitMemory}{A boolean (True/False) indicating whether to limit the memory use or not. Default is true. If set to true, 5000 datapoints are randomly sampled from each dataset under comparison for inference.}

\item{opt_method}{A string specifying the optimization method to be used for hyperparameter estimation. Current options are: 'L-BFGS-B' and 'BFGS'. Default is set to 'L-BFGS-B' and is recommended.}
}
\value{
a list containing :
 \itemize{
  \item muDiff - A vector of pointwise difference between the predictions from the two datasets \code{(mu2- mu1)}
  \item mu1 - A vector of test prediction for first data set
  \item mu2 - A vector of test prediction for second data set
  \item band - A vector of the allowed statistical difference between functions at testpoints in testset
  \item confLevel - A numeric representing the statistical significance level for constructing the band
  \item testset - A matrix of test points to compare the functions
  \item estimatedParams - A list of estimated hyperparameters for GP
}
}
\description{
Function comparison using Gaussian Process and Hypothesis testing
}
