% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnFit.R
\name{KnnPcurve}
\alias{KnnPcurve}
\title{KNN : Fit}
\usage{
KnnPcurve(data, xCol, yCol, subsetSelection = FALSE)
}
\arguments{
\item{data}{a dataframe or a matrix, to be used in modelling}

\item{xCol}{a vector or numeric values stating the column number of features}

\item{yCol}{a numerical or a vector value stating the column number of target}

\item{subsetSelection}{a boolean, default value is FALSE, if TRUE returns the best beature combination}
}
\value{
a list containing :
        1) data
        2) column number of features
        3) column number of target
        4) best K
}
\description{
The function models the powercurve using KNN, against supplied arguments
}
