% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnPredict.R
\name{KnnPredict}
\alias{KnnPredict}
\title{KNN : Predict}
\usage{
KnnPredict(knnMdl, testData)
}
\arguments{
\item{testData}{a data frame or matrix, to compute the predictions}

\item{KnnMdl}{a list containing:
 \itemize{
  \item knnMdl$data - The data set provided by user
  \item knnMdl$xCol - The column number of features provided by user or the best subset column number
  \item knnMdl$yCol - The column number of target provided by user
  \item knn$bestK - The best k nearest neighbor calculated using the function KnnFit
}}
}
\value{
a list containing :
 \itemize{
  \item data - The data set provided by user
  \item xCol - The column number of features provided by user or the best subset column number
  \item yCol - The column number of target provided by user
  \item prediction - The test data prediction using model generated by KnnFit
}
}
\description{
The function can be used to make prediction on test data using trained model
}
