% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnnUpdate.R
\name{KnnUpdate}
\alias{KnnUpdate}
\title{KNN : Update}
\usage{
KnnUpdate(knnMdl, newData)
}
\arguments{
\item{newData}{a dataframe or a matrix, to be used for updating the model}

\item{KnnMdl}{a list containing:
 \itemize{
  \item knnMdl$data - The data set provided by user
  \item knnMdl$xCol - The column number of features provided by user or the best subset column number
  \item knnMdl$yCol - The column number of target provided by user
  \item knn$bestK - The best k nearest neighbor calculated using the function KnnFit
}}
}
\value{
a list containing :
 \itemize{
  \item data - The updated data using old data set and new data
  \item xCol - The column number of features provided by user or the best subset column number
  \item yCol - The column number of target provided by user
  \item bestK - The best k nearest neighbor calculated for the new data using user specified features and target
}
}
\description{
The function can be used to update KNN model when new data is provided
}
