% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplinePCFit.R
\name{SplinePCFit}
\alias{SplinePCFit}
\title{Smoothing spline Anova method}
\usage{
SplinePCFit(data, xCol, yCol, testP)
}
\arguments{
\item{data}{a matrix or dataframe to be used in modelling}

\item{xCol}{a numeric or vector stating the column number of covariates}

\item{yCol}{A numeric value stating the column number of target}

\item{testX}{a matrix or dataframe, to be used in computing the predictions}
}
\value{
a vector or numeric predictions on user provided test data
}
\description{
Smoothing spline Anova method
}
