% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplinePCFit.R
\name{SplinePCFit}
\alias{SplinePCFit}
\title{Smoothing spline Anova method}
\usage{
SplinePCFit(data, xCol, yCol, testX, modelFormula = NULL)
}
\arguments{
\item{data}{a matrix or dataframe to be used in modelling}

\item{xCol}{a numeric or vector stating the column number of feature covariates}

\item{yCol}{a numeric value stating the column number of target}

\item{testX}{a matrix or dataframe, to be used in computing the predictions}

\item{modelFormula}{default is NULL else a model formula specifying target and features.Please refer 'gss' package documentation for more details}
}
\value{
a vector or numeric predictions on user provided test data
}
\description{
Smoothing spline Anova method
}
\examples{

data = data1[c(1:100),]
xCol = 2
yCol = 7
testX = data1[c(101:110), ]
Spline_prediction = SplinePCFit(data, xCol, yCol, testX)

}
