% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempGP.R
\name{tempGP}
\alias{tempGP}
\title{temporal Gaussian process}
\usage{
tempGP(trainX, trainY)
}
\arguments{
\item{trainX}{A matrix with each column corresponding to one input variable.}

\item{trainY}{A vector with each element corresponding to the output at the corresponding row of \code{trainX}.}
}
\value{
An object of class \code{tempGP} with the following attributes:
\itemize{
\item trainX
\item trainY
\item thinningNumber
\item estimatedParams
\item llval
\item gradval
}
}
\description{
A Gaussian process based power curve model which explicitly models the temporal aspect of the power curve.
}
